+++ R/EMstep.group.R
@@ -1,4 +1,4 @@
-EM.group <- function(pars, constrain, Ls, Data, PrepList, list, Theta, DERIV, nloptr_args)
+EM.group <- function(pars, constrain, Ls, Data, PrepList, list, Theta, DERIV, solnp_args)
 {
     verbose <- list$verbose
     nfact <- list$nfact
@@ -197,7 +197,7 @@ EM.group <- function(pars, constrain, Ls, Data, PrepList, list, Theta, DERIV, nl
                               PrepList=PrepList, L=L, UBOUND=UBOUND, LBOUND=LBOUND, Moptim=Moptim,
                               BFACTOR=BFACTOR, nfact=nfact, Thetabetween=Thetabetween, 
                               rlist=rlist, constrain=constrain, DERIV=DERIV, Mrate=Mrate, 
-                              TOL=list$MSTEPTOL, nloptr_args=nloptr_args)
+                              TOL=list$MSTEPTOL, solnp_args=solnp_args)
             if(list$accelerate && Mrate > .01 && cycles %% 3 == 0L){
                 dX2 <- preMstep.longpars - preMstep.longpars2
                 dX <- longpars - preMstep.longpars
@@ -301,7 +301,7 @@ Estep.bfactor <- function(pars, tabdata, freq, Theta, prior, Prior, Priorbetween
 }
 
 Mstep <- function(pars, est, longpars, ngroups, J, gTheta, itemloc, PrepList, L, ANY.PRIOR,
-                  UBOUND, LBOUND, constrain, DERIV, Prior, rlist, CUSTOM.IND, nloptr_args,
+                  UBOUND, LBOUND, constrain, DERIV, Prior, rlist, CUSTOM.IND, solnp_args,
                   SLOW.IND, groupest, BFACTOR, nfact, Thetabetween, Moptim, Mrate, TOL){
     p <- longpars[est]
     if(length(p)){
@@ -343,20 +343,17 @@ Mstep <- function(pars, est, longpars, ngroups, J, gTheta, itemloc, PrepList, L,
                                 J=J, gTheta=gTheta, PrepList=PrepList, L=L,  ANY.PRIOR=ANY.PRIOR,
                                 constrain=constrain, LBOUND=LBOUND, UBOUND=UBOUND, SLOW.IND=SLOW.IND,
                                 itemloc=itemloc, DERIV=DERIV, rlist=rlist, TOL=TOL))
-        } else if(Moptim %in% c('nloptr', 'nloptr_no_grad')){
-            if(Moptim == 'nloptr_no_grad') eval_grad_f <- NULL
-            else eval_grad_f <- Mstep.grad_alt
+        } else if(Moptim == 'solnp'){
             optim_args <- list(CUSTOM.IND=CUSTOM.IND, est=est, longpars=longpars, pars=pars,
                                ngroups=ngroups, J=J, gTheta=gTheta, PrepList=PrepList, L=L,
                                ANY.PRIOR=ANY.PRIOR, constrain=constrain, LBOUND=LBOUND,
                                UBOUND=UBOUND, SLOW.IND=SLOW.IND, itemloc=itemloc, DERIV=DERIV, 
                                rlist=rlist)
-            opt <- try(nloptr(p, eval_f=Mstep.LL_alt, eval_grad_f=eval_grad_f,
-                          lb=nloptr_args$lb, ub=nloptr_args$ub, eval_g_ineq=nloptr_args$eval_g_ineq,
-                          eval_jac_g_ineq=nloptr_args$eval_jac_g_ineq, 
-                          eval_g_eq=nloptr_args$eval_g_eq, eval_jac_g_eq=nloptr_args$eval_jac_g_eq,
-                          opts = nloptr_args$opts, optim_args=optim_args), silent=TRUE)
-            if(!is(opt, 'try-error')) opt$par <- opt$solution
+            opt <- try(solnp(p, Mstep.LL_alt, eqfun = solnp_args$eqfun, eqB = solnp_args$eqB, 
+                             ineqfun = solnp_args$ineqfun, ineqLB = solnp_args$ineqLB, 
+                             ineqUB = solnp_args$ineqUB, LB = solnp_args$LB, UB = solnp_args$UB,
+                             control=solnp_args$control, optim_args=optim_args), silent=TRUE)
+            if(!is(opt, 'try-error')) opt$par <- opt$pars
         } else {
             stop('M-step optimizer not supported')
         }
@@ -477,16 +474,6 @@ Mstep.grad <- function(p, est, longpars, pars, ngroups, J, gTheta, PrepList, L, 
         grad <- g
     }
     return(-grad[est])
-}
-
-Mstep.grad_alt <- function(x0, optim_args){
-    return(Mstep.grad(p=x0, est=optim_args$est, longpars=optim_args$longpars, pars=optim_args$pars, 
-                    ngroups=optim_args$ngroups, J=optim_args$J, gTheta=optim_args$gTheta, 
-                    PrepList=optim_args$PrepList, L=optim_args$L, CUSTOM.IND=optim_args$CUSTOM.IND, 
-                    SLOW.IND=optim_args$SLOW.IND, 
-                    constrain=optim_args$constrain, LBOUND=optim_args$LBOUND, 
-                    UBOUND=optim_args$UBOUND, itemloc=optim_args$itemloc,
-                    DERIV=optim_args$DERIV, rlist=optim_args$rlist, ANY.PRIOR=optim_args$ANY.PRIOR))
 }
 
 Mstep.NR <- function(p, est, longpars, pars, ngroups, J, gTheta, PrepList, L,  ANY.PRIOR,